Update README.md
Stroke has led to a lot of deaths, and it is of high necessity to put it into consideration as many deaths have occurred through this health issue. Early prediction is essential, and this project is to help with the use of machine learning to predict stroke detection at an early stage so must administer early treatment and prevent death. In the focus of this project, we use various algorithms to develop skills and knowledge in various algorithms which will help achieve this goal. With the use of various algorithms, we will be able to view which algorithm works best for this kind of type of prediction. We have been able to use a comprehensive and robust dataset which has essential factors to use for this prediction.
This is a classification problem, therefore, various algorithms which address this type of problem were used in the likes of logistic regression, K-Nearest Neighbours, Naive Bayes, decision trees, and random forests, SVC. Performance metrics (accuracy, precision, recall, f1- score) were used to investigate the performance of each model.